Factorize childs in choices
