+++ app/javascript/components/AdvancedComplexDatationSearch/components/ChoiceSetSearch.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useMemo } from 'react';
 import ReactSelect from 'react-select';
 import Translations from '../../Translations/components/Translations';
+import ChildChoicesElement from '../../AdvancedSearchShared/ChildChoicesElement';
 
 const ChoiceSetSearch = (props) => {
   const {
@@ -61,22 +62,11 @@ const ChoiceSetSearch = (props) => {
   }
 
   function renderChildChoicesElement() {
-    const childChoices = [
-      { value: true, label: Translations.messages['advanced_searches.fields.choice_set_search_field.child_choices_activated_yes'] },
-      { value: false, label: Translations.messages['advanced_searches.fields.choice_set_search_field.child_choices_activated_no'] },
-    ];
-    const isChildChoicesActivated = defaultValues["child_choices_activated"] && defaultValues["child_choices_activated"] === "true";
-
-    const defaultValue = childChoices.find(
-      (item) => item.value === isChildChoicesActivated,
-    );
-
     return (
-      <ReactSelect name={`advanced_search[criteria][${fieldUuid}][${itemId}][child_choices_activated]`}
-        options={childChoices}
-        classNamePrefix="select"
-        placeholder={Translations.messages['advanced_searches.fields.choice_set_search_field.child_choices_activated_placeholder']}
-        defaultValue={defaultValue}
+      <ChildChoicesElement
+        fieldUuid={fieldUuid}
+        itemId={itemId}
+        defaultValues={defaultValues}
       />
     );
   }
